refactor: split the tokenizer replacer in separate files
